[miniforge] no fail on existing installation (#167)
Fixes #166 

Note that this does not handle the case where the existing version is
not the version requested. I started to address it, but it is made more
complicated by the fact that mambaforge versions don't seem to keep up
with official conda releases, so I didn't venture further down that path
(i.e. if someone comes with an image with a fresh conda install, we
would almost certainly be down-grading it here, which might not be
desired).

I also updated the devcontainer spec of this repo because it was using
an outdated `docker-from-docker` feature, which in turn made the tests
in the tests break when running in the devcontainer.